\name{timeDF-class}
\alias{timeDF-class}
\title{
timeDF S3 class
}
\description{
timeDF object stores definitions of periods.
}
\details{
  timeDF object stores records with time information. The column to hold
  time information can be specified as "time_var" attribute.
}
\seealso{
  \code{\link{as.timeDF}}
  \code{\link{timeDF-package}}
}
